backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
python-multipart==0.0.17
redis==5.2.0
celery==5.4.0
httpx==0.28.0
numpy==2.1.3
scikit-learn==1.5.2
Pillow==11.0.0
python-dotenv==1.0.1
